Redirect FAQ and 1liners to Eric Pement's site
In November 2022 the sed.sf.net website was removed and we lost
the write access to update it. Read the details in:
https://github.com/aureliojargas/sed.sf.net/issues/11

Since Eric Pement already have those documents in his website, doing
the redirect we avoid the duplication and workaround the access
problem.

Doing a soft "temp" redirect so the Sourceforge address is still
considered the official one, as stated in Eric's documents.